# src/zenml/zen_server/openapi3_0.yaml
openapi: 3.0.2
info:
  title: ZenML
  version: 0.xx.0
  description: >-
    The Zen API allows access to ZenML programmatically. These docs DO NOT represent the status quo but instead are the Design that is being aimed at.
...
paths:
...
  /pipelines:
    get:
      summary: List pipelines
      tags:
        - pipeline
      description: |-
        Get a list of pipelines.
      operationId: get_pipelines      
      parameters:      
        - in: query
          name: project_name
          required: false
          schema:
            title: Project Name
            type: string
          description: >-
            Pipelines are unique and defined within the project scope. It is
            expected that this endpoint will be used at project scope most of 
            the time.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: List of Pipelines
                type: array
                items:
                  $ref: '#/components/schemas/Pipeline'
        '401': 
          $ref: '#/components/responses/401'
        '404': 
          $ref: '#/components/responses/404'
        '422': 
          $ref: '#/components/responses/422'
      security:
        - HTTPBasic: []
  /pipelines/{pipeline_id}:
    get:
      summary: Get a specific pipeline
      tags:
        - pipeline
      description: |-
        Get a specific pipeline using its unique id.
      operationId: get_pipeline_by_id
      parameters:
        - $ref: '#/components/parameters/pipelineId'
      responses:
        '200':
          description: Returns the requested pipeline.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pipeline'
        '401': 
          $ref: '#/components/responses/401'
        '404': 
          $ref: '#/components/responses/404'
        '422': 
          $ref: '#/components/responses/422'

      security:
        - HTTPBasic: []
    put:
      summary: Update a specific pipeline
      tags:
        - pipeline
      description: |-
        Updates the attribute on a pipeline.
      operationId: update_pipeline_by_id
      parameters:
        - $ref: '#/components/parameters/pipelineId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pipeline'
        required: true
      responses:
        '200':
          description: Returns the updated pipeline.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pipeline'
        '401': 
          $ref: '#/components/responses/401'
        '404': 
          $ref: '#/components/responses/404'
        '422': 
          $ref: '#/components/responses/422'
      security:
        - HTTPBasic: []
    delete:
      summary: Delete a specific pipeline
      tags:
        - pipeline
      description: |-
        Deletes a pipeline.
      operationId: delete_pipeline_by_id
      parameters:
        - $ref: '#/components/parameters/pipelineId'
      responses:
        '200': 
          $ref: '#/components/responses/200'
        '401': 
          $ref: '#/components/responses/401'
        '404': 
          $ref: '#/components/responses/404'
        '422': 
          $ref: '#/components/responses/422'
      security:
        - HTTPBasic: []
...
  /runs:
    get:
      summary: Get all pipeline runs
      tags:
        - run
      description: |-
        Get pipeline runs according to query filters.
      operationId: get_pipeline_runs
      parameters:    
        - in: query
          name: project_name
          required: false
          schema:
            title: Project Name
            type: string
        - name: stack_id
          in: query
          required: false
          schema:
            title: Stack Id
            type: string
        - name: pipeline_id
          in: query
          required: false
          schema:
            title: Pipeline Id
            type: string
        - name: trigger_id
          in: query
          required: false
          schema:
            title: Trigger Id
            type: string
      responses:
        '200':
          description: Returns the requested pipeline run.
          content:
            application/json:
              schema:
                title: List all pipeline runs
                type: array
                items:
                  $ref: '#/components/schemas/Run'
        '401': 
          $ref: '#/components/responses/401'
        '404': 
          $ref: '#/components/responses/404'
        '422': 
          $ref: '#/components/responses/422'

      security:
        - HTTPBasic: []
...
  /projects/{project_name}/pipelines:
    get:
      summary: Get pipelines that are defined within a project
      tags:
        - project
      description: |-
        Returns all pipelines within the project.
      operationId: get_pipelines_in_project
      parameters:
        - $ref: '#/components/parameters/projectName'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: List of Stack Components in Project
                type: array
                items:
                  $ref: '#/components/schemas/Pipeline'
        '401': 
          $ref: '#/components/responses/401'
        '404': 
          $ref: '#/components/responses/404'
        '422': 
          $ref: '#/components/responses/422'
      security:
        - HTTPBasic: []
    post:
      summary: Create a pipeline within a project
      tags:
        - project
      description: |-
        Creates a pipeline based on the requestBody.
      operationId: create_pipeline
      parameters:
        - $ref: '#/components/parameters/projectName'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pipeline'
        required: true
      responses:
        '200':
          description: Returns the created Pipeline.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pipeline'
        '401': 
          $ref: '#/components/responses/401'
        '409': 
          $ref: '#/components/responses/409'
        '422': 
          $ref: '#/components/responses/422'
      security:
        - HTTPBasic: []
...
  /metadata-config:
    get:
      summary: Metadata Config
      tags:
        - metadata-config
      description: |-
        Metadata config that allows the Orchestrator to bypass the API and connect directly to the
        Metadata Database tables to read and write run data.
      operationId: get_metadata_config
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: Metadata Config
                type: object
        '401': 
          $ref: '#/components/responses/401'
        '404': 
          $ref: '#/components/responses/404'
        '422': 
          $ref: '#/components/responses/422'
      security:
        - HTTPBasic: [] 
components:
...
  schemas:
...
    Pipeline:
      title: Pipeline
      required:
        - name
        - id
      type: object
      properties:
        id:
          title: id
          type: string
          readOnly: True
          example: 9eaa2d6
        name:
          title: Name
          type: string
          example: cat_detection_training_pipeline
        description:
          title: Description
          type: string
          example: >-
            Pipeline to create a cat detector.
        status:
          $ref: '#/components/schemas/PreviousRunsStatus'
        owner:
          title: Owner
          type: object
          properties:
            id:
              title: id
              type: string
              readOnly: True
            name:
              title: Name
              type: string   
          readOnly: True
          description: >-
            User that created the trigger.
          example:
            {name: htahir, id: w813dw}
        created_at:
          title: Created at
          type: string
          format: date-time
          readOnly: True
          description: >-
            Time when this pipeline was first created.
          example: 2022-08-12T07:12:45.931Z
      description: |-
        A Pipeline is the blueprint for a collection of functions (or steps) 
        that are meant to fulfill one overarching task. 
        The flow of information through these functions is defined by a
        dag (directed acyclic graph).
...
    Run:
      title: Run
      required:
        - id
        - name
        - pipeline
        - stack
        - runtime_configuration
        - user_id
        - project_name
      type: object
      properties:
        id:
          title: id
          type: string
          readOnly: True
          example: 68be3034
        name:
          title: Name
          type: string
          example: cat_detection_training_pipeline_run_4921
        zenml_version:
          title: Zenml Version
          type: string
          description: >-
            The ZenML Version that the pipeline was run on.
          example: 0.12.0
        duration:
          title: duration
          type: string
          description: >-
            The dureation of the last run.
          readOnly: True
        git_sha:
          title: Git Sha
          type: string
          description: >-
            The GitSha of the code repository this pipeline was run with
          example: 73841b4
        pipeline:
          type: object
          properties:
            id:
              title: id
              type: string
              readOnly: True
            name:
              title: Name
              type: string   
          description: >-
            The pipeline that the run is an instance of.
          example:
            {name: cat_detection_training_pipeline, id: 9eaa2d6}
        stack:
          type: object
          properties:
            id:
              title: id
              type: string
              readOnly: True
            name:
              title: Name
              type: string
          description: >-
            The stack that was used for this run.
          example:
            {name: prod_stack, id: 9lg2jk3}
        trigger:
          type: object
          properties:
            id:
              title: id
              type: string
              readOnly: True
            name:
              title: Name
              type: string
          description: >-
            The trigger that was used for this run.
          example:
            {name: weekly_prod_trigger, id: 381jd1k}
        status:
          type: string
          enum:
            - Running
            - Failed
            - Finished
          readOnly: True
          example: Finished
        owner:
          title: Owner
          type: object
          properties:
            id:
              title: id
              type: string
              readOnly: True
            name:
              title: Name
              type: string   
          readOnly: True
          description: >-
            User that owns this run.
          example:
            {name: htahir, id: w813dw}
        project_name:
          title: Project Name
          type: string
          example: demo_prj
        created_at:
          title: Run started at 
          type: string
          format: date-time
          readOnly: True
          description: >-
            Run started at.
          example: 2022-08-12T07:12:45.931Z
      description: |-
        A single Pipeline run. 
...
    PreviousRunsStatus:
      title: Status' of the Previous Run
      type: object
      properties:
        1:
          $ref: '#/components/schemas/RunStatus'
        2:
          $ref: '#/components/schemas/RunStatus'
        3:
          $ref: '#/components/schemas/RunStatus'
    RunStatus:
      title: Run status
      type: object
      properties:
        status: 
          type: string
          enum:
            - Running
            - Failed
            - Finished
          readOnly: True
          example: Finished
        run:
          type: object
          properties:
            id:
              title: id
              type: string
              readOnly: True
            name:
              title: Name
              type: string
          description: >-
            The run.
...
    ErrorModel:
      title: ErrorModel
      type: object
      properties:
        detail:
          title: Detail
      description: Base class for error responses.
    ValidationError:
      title: ValidationError
      required:
        - loc
        - msg
        - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            anyOf:
              - type: string
              - type: integer
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
  #-------------------------------
  # Reusable operation parameters
  #-------------------------------
  parameters:
    pipelineId:
      name: pipeline_id
      in: path
      required: true
      schema:
        title: Pipeline ID
        type: string
...
    projectName:
      name: project_name
      in: path
      required: true
      schema:
        title: Project Name
        type: string
...
  responses:
    '200':
      description: Successful Response
    '401':
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorModel'
    '404':
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorModel'        
    '409':
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorModel'
    '422':
      description: Validation Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/HTTPValidationError'
...
  securitySchemes:
    HTTPBasic:
      type: http
      scheme: basic